Created namespace Analyze. Split Analyze into Curve and Surface.
