--- src/GShark.Test.XUnit/Operation/CheckTests.cs
@@ -1,6 +1,0 @@
-﻿namespace GShark.Test.XUnit.Operation
-{
-    public class CheckTests
-    {
-    }
-}
